move to a useStorybookDarkMode hook
